requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.4
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
environs==11.0.0
gunicorn==21.2.0
idna==3.6
marshmallow==3.21.1
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
